requirements.txt
fastapi
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
# pymysql==1.1.0  # MySQL only
pytz==2025.2
gunicorn
supabase>=2.3.4
psycopg2-binary==2.9.9
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6